requirements.txt
discord>=2.3.2
python-dotenv>=1.0.1
moviepy>=1.0.3
assemblyai>=0.26.0